Added cpp's and documentation for classes
Changed the name of FitnessLog to just Log.

Converted every class to have a header file and a .cpp file.

Used Doxygen style code comments in the header files for all the classes that make up MealLog and WorkoutLog.

Removed the error handling components. Plan on adding it back but need to start from the ground up with it.